+++ src/util/index.ts
@@ -32,7 +32,8 @@ import {
 import invariant from '../util/invariant';
 import { getConfigDirectoryPath } from './config/manage';
 import { sha256 } from './createHash';
-import { convertTestResultsToTableRow, getHeaderForTable } from './exportToFile';
+import { convertTestResultsToTableRow } from './exportToFile';
+import { getHeaderForTable } from './exportToFile/getHeaderForTable';
 import { maybeLoadFromExternalFile } from './file';
 import { isJavascriptFile } from './fileExtensions';
 import { getNunjucksEngine } from './templates';
